Conflicts:
	build/cocos2d_libs.xcodeproj/project.pbxproj
	cocos/scripting/lua/bindings/CCLuaStack.cpp
	cocos/scripting/lua/script/Cocos2d.lua
	cocos/scripting/lua/script/Cocos2dConstants.lua

+++ cocos/2d/ccUTF8.h
@@ -80,7 +80,7 @@ CC_DLL unsigned short* cc_utf8_to_utf16(const char* str_old, int length = -1, in
  **/
 CC_DLL char *
 cc_utf16_to_utf8 (const unsigned short  *str,
-                  long             len,
+                  int             len,
                   long            *items_read,
                   long            *items_written);
